Expanding leaks fixed
