+++ sources/parser/expand_utils3.c
