+++ sources/parser/handle_cmd_args.c
@@ -6,7 +6,7 @@
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 14:26:25 by henbuska          #+#    #+#             */
-/*   Updated: 2024/12/20 11:34:27 by nzharkev         ###   ########.fr       */
+/*   Updated: 2024/12/20 17:48:37 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -68,17 +68,15 @@ int	init_args_array(t_cmd *cmd, int i)
 	return (0);
 }
 
-int export_args(t_shell *mini, t_cmd *cmd, t_expand *arg, int i)
+int export_args(t_cmd *cmd, t_expand *arg, int i)
 {
-		arg->sgl = 0;
+	arg->sgl = 0;
 	arg->dbl = 0;
 	arg->i = i;
 	arg->start = i;
-	(void)mini;
 	what_quote(cmd->segment, arg);
 	while (cmd->segment[arg->i])
 	{
-		// printf("arg->value: {%s}\nlen: %d\n", arg->value, (int)ft_strlen(arg->value));
 		if (ft_isspace(cmd->segment[arg->i]) && !arg->sgl && !arg->dbl)
 			break ;
 		if (!arg->sgl && !arg->dbl && (cmd->segment[arg->i] == '\'' || cmd->segment[arg->i] == '"'))
@@ -96,7 +94,6 @@ int export_args(t_shell *mini, t_cmd *cmd, t_expand *arg, int i)
 		else if (add_char(cmd->segment, arg))
 			return (free(arg->value), -1);
 	}
-	// printf("value: {%s}\n", arg->value);
 	arg->len = ft_strlen(arg->value);
 	return (arg->i);
 }
@@ -105,7 +102,6 @@ int only_redirect(char *str, int i)
 {
 	int n;
 	n = i;
-	// printf("str[%d]: {%c}\n", n, str[n]);
 	while (str[n] && str[n] != '=')
 		n++;
 	if (!str[n])
@@ -119,24 +115,22 @@ int only_redirect(char *str, int i)
 
 int	handle_arg(t_shell *mini, t_cmd *cmd, int i, t_expand *arg, int *arg_index)
 {
-	// printf("cmd->segment[%d]: {%c}\n", i, cmd->segment[i]);
 	if (ft_strcmp(cmd->command, "export") == 0 && only_redirect(cmd->segment, i))
-		i = export_args(mini, cmd, arg, i);
+		i = export_args(cmd, arg, i);
 	if (cmd->segment[i] == '\'' || cmd->segment[i] == '"')
 		i = arg_in_quotes(mini, cmd->segment, i, arg);
 	else
 		i = arg_no_quotes(mini, cmd, arg, i);
 	if (i == -1)
 		return (-1);
-	// printf("value: {%s}\nsegment: {%c}\n", arg->value, cmd->segment[i]);
 	if (!arg->value || append_to_array(cmd, arg->value, arg->len, arg_index) == -1)
 	{
 		free(arg->value);
 		ft_free_array(cmd->args);
 		return (-1);
 	}
-	// free(arg->value);
-	arg->value = ft_strdup("");
+	free(arg->value);
+	// arg->value = ft_strdup("");
 	i = skip_whitespace(cmd->segment, i);
 	return (i);
 }
@@ -151,6 +145,7 @@ int	handle_cmd_args(t_shell *mini, t_cmd *cmd, int i)
 	arg_index = 1;
 	i = skip_whitespace(cmd->segment, i);
 	the_arg(&arg, i);
+	free(arg.value);
 	while (cmd->segment[i] && arg_index < cmd->args_count + 1)
 	{
 		if (is_redirection(cmd, i))
